@@ -64,13 +64,13 @@
     "solid-js": "^1.0.0"
   },
   "devDependencies": {
-    "@vitest/ui": "^0.21.1",
+    "@vitest/ui": "^0.26.2",
     "c8": "^7.12.0",
     "npm-run-all": "^4.1.5",
-    "release-it": "^15.3.0",
-    "solid-js": "^1.4.8",
-    "typescript": "^4.7.4",
-    "vite": "^2.9.15",
+    "release-it": "^15.6.0",
+    "solid-js": "^1.6.6",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.3",
     "vitest": "^0.21.1"
   },
   "engines": {
