@@ -3,7 +3,7 @@
 import path from "path";
 import { defineConfig } from "vite";
 
-module.exports = defineConfig({
+export default defineConfig({
   build: {
     target: "esnext",
     lib: {
